Update DrawingFile Controller

+++ DrawingRegisterWeb/Controllers/DrawingFilesController.cs
@@ -2,43 +2,91 @@
 using Aspose.Pdf.Devices;
 using DrawingRegisterWeb.Data;
 using DrawingRegisterWeb.Models;
-using DrawingRegisterWeb.ViewModels;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Hosting;
 
 namespace DrawingRegisterWeb.Controllers
 {
+	// DrawingRegister - Creates a unique identifier
+
+	// DrawingRegisterUsers - Creates a relationship with DrawingRegisters and AspNetUsers
+
+	// ProjectState - Defines states for projects, creates a relationship with projects and DrawingRegisters,
+	//				  restricts access to DrawingRegisterUsers that share the same DrawingRegister as the current user
+
+	// Project - Holds main data about user project. Seperates and group drawings, documentation.
+	//			 Creates the relationship for drawings, documentation and layouts to ProjectStates
+
+	// Drawings - One of the building blocks of the project.
+	//			  Holds main data about drawing files that belong to area/equipment/construction and so on.
+	//			  Creates the relationship between DrawingFiles and project
+
+	[Authorize(Roles = $"{ConstData.Role_Admin_Name},{ConstData.Role_Engr_Name}")]
 	public class DrawingFilesController : Controller
 	{
 		private readonly DrawingRegisterContext _context;
 		private readonly IWebHostEnvironment _hostEnvironment;
+		private readonly UserManager<IdentityUser> _userManager;
 
-		public DrawingFilesController(DrawingRegisterContext context, IWebHostEnvironment hostEnvironment)
+		public DrawingFilesController(
+			DrawingRegisterContext context,
+			IWebHostEnvironment hostEnvironment,
+			UserManager<IdentityUser> userManager)
 		{
 			_context = context;
 			_hostEnvironment = hostEnvironment;
+			_userManager = userManager;
 		}
 
-		// POST: DrawingFiles/Create
+
+
+
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> Create(string drawingId, List<IFormFile> files, bool automaticRevision = false)
+		public async Task<IActionResult> Create(int drawingId, List<IFormFile> files, bool automaticRevision = false)
 		{
+			var user = await _userManager.GetUserAsync(User);
+			var drawingRegisterUser = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(dr => dr.UserId == user.Id);
+
+			var drawing = await _context.Drawing
+				.Include(d => d.Project)
+				.Include(d => d.Project.ProjectState)
+				.Where(d => d.Project.ProjectState!.DrawingRegisterId == drawingRegisterUser!.DrawingRegisterId)
+				.FirstOrDefaultAsync(d => d.Id == drawingId);
+
+			// Check if drawing exists and ensure that user access to drawing that is only in his DrawingRegister
+			if (drawing == null)
+			{
+				return NotFound();
+			}
+
+			// Make sure that only administrator can create drawing files if ProjectState is Defined, Completed or Canceled
+			if (drawingRegisterUser!.Role != ConstData.Role_Admin_Name)
+			{
+				if (drawing!.Project.ProjectState!.Name == ConstData.State_Defined ||
+					drawing!.Project.ProjectState!.Name == ConstData.State_Completed ||
+					drawing!.Project.ProjectState!.Name == ConstData.State_Canceled)
+				{
+					TempData["NoUpload"] = $"Only the administrator has the ability to upload files to this drawing " +
+						$"if project state is set to {drawing!.Project.ProjectState!.Name}.";
+
+					return RedirectToAction("Details", "Drawings", new { id = drawingId });
+				}
+			}
+
 			if (ModelState.IsValid)
 			{
 				foreach (var file in files)
 				{
-					var drawing = _context.Drawing.FirstOrDefault(p => p.Id == Int32.Parse(drawingId));
-
+					// Get new file path, Guid name and extension
 					string wwwRootPath = _hostEnvironment.WebRootPath;
-
 					string fileName = Guid.NewGuid().ToString();
 					var uploads = Path.Combine(wwwRootPath, @"Files\Drawings");
 					var extension = Path.GetExtension(file.FileName)!.ToLower();
 
+					// Create DrawingFile - Drawing file keeps data about uploaded file
 					var drawingFile = new DrawingFile()
 					{
 						DrawingId = drawing!.Id,
@@ -48,22 +96,23 @@ namespace DrawingRegisterWeb.Controllers
 						FileUrl = @"\Files\Drawings\" + fileName + extension
 					};
 
+					// if automatic Revision checked - add revision to drawing file from file name
 					if (automaticRevision)
 					{
-						if (file.FileName != null && file.FileName.Contains("_"))
+						if (file.FileName != null && file.FileName.Contains('_'))
 						{
-							int fileNameEndIndex = drawingFile.FileName.LastIndexOf("_");
-							drawingFile.Revision = Path.GetFileNameWithoutExtension(file.FileName)!.Substring(fileNameEndIndex+1);
+							int fileNameEndIndex = drawingFile.FileName.LastIndexOf('_');
+							drawingFile.Revision = Path.GetFileNameWithoutExtension(file.FileName)!.Substring(fileNameEndIndex + 1);
 						
 						}
 					}
-
 
 					using (var fileStream = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
 					{
 						file.CopyTo(fileStream);
 					}
 
+					// If file is pdf type - create files thumbnail
 					if(drawingFile.FileType.ToLower() == "pdf")
 					{
 						var pdfDocument = new Document(Path.Combine(uploads, fileName + extension));
@@ -90,7 +139,9 @@ namespace DrawingRegisterWeb.Controllers
 			return RedirectToAction("Details", "Drawings", new { id = drawingId });
 		}
 
-		// GET: DrawingFiles/Edit
+
+
+
 		public async Task<IActionResult> Edit(int? id)
 		{
 			if (id == null || _context.DrawingFile == null)
@@ -98,28 +149,77 @@ namespace DrawingRegisterWeb.Controllers
 				return NotFound();
 			}
 
-			var drawingFile = await _context.DrawingFile.FindAsync(id);
+			var user = await _userManager.GetUserAsync(User);
+			var drawingRegisterUser = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(dr => dr.UserId == user.Id);
+			var drawingFile = await _context.DrawingFile
+				.Include(d => d.Drawing)
+				.Include(d => d.Drawing.Project)
+				.Include(d => d.Drawing.Project.ProjectState)
+				.Where(d => d.Drawing.Project.ProjectState!.DrawingRegisterId == drawingRegisterUser!.DrawingRegisterId)
+				.FirstOrDefaultAsync(d => d.Id == id);
 
+			// Check if drawing file exists and ensure that user access to drawing that is only in his DrawingRegister
 			if (drawingFile == null)
 			{
 				return NotFound();
 			}
-			ViewData["DrawingId"] = new SelectList(_context.Drawing, "Id", "DrawingNumber", drawingFile.DrawingId);
+
+			// Return aproved DrawingId
+			ViewData["DrawingId"] = drawingFile.DrawingId;
 			return View(drawingFile);
 		}
 
-		// POST: DrawingFiles/Edit
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> Edit(int id, DrawingFile drawingFile)
+		public async Task<IActionResult> Edit(int? id, DrawingFile drawingFile)
 		{
-			if (id != drawingFile.Id)
+			var user = await _userManager.GetUserAsync(User);
+			var drawingRegisterUser = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(dr => dr.UserId == user.Id);
+
+			// Check if drawingFile exsist and id is correct
+			if (id == null || drawingFile == null || id != drawingFile.Id)
 			{
 				return NotFound();
 			}
 
+			var drawingFileBeforeEdit = await _context.DrawingFile
+				.Include(d => d.Drawing)
+				.Include(d => d.Drawing.Project)
+				.Include(d => d.Drawing.Project.ProjectState)
+				.Where(d => d.Drawing.Project.ProjectState!.DrawingRegisterId == drawingRegisterUser!.DrawingRegisterId)
+				.FirstOrDefaultAsync(d => d.Id == id);
+
+			// Check if DrawingFile belongs to current users DrawingRegister
+			if(drawingFileBeforeEdit == null) 
+			{ 
+				return NotFound(); 
+			}
+
+			// Make sure that only administrator can edit if ProjectState is Defined, Completed or Canceled
+			if (drawingRegisterUser!.Role != ConstData.Role_Admin_Name)
+			{
+				if (drawingFileBeforeEdit!.Drawing.Project.ProjectState!.Name == ConstData.State_Defined ||
+					drawingFileBeforeEdit!.Drawing.Project.ProjectState!.Name == ConstData.State_Completed ||
+					drawingFileBeforeEdit!.Drawing.Project.ProjectState!.Name == ConstData.State_Canceled)
+				{
+					ModelState.AddModelError("NoEdit",
+						$"Only the administrator has the ability to edit this drawing file " +
+						$"if project state is set to {drawingFileBeforeEdit!.Drawing.Project.ProjectState!.Name}.");
+				}
+			}
+
+			// Prevent from white space drawing number and revision
+			if (drawingFile.FileName == null || drawingFile.Revision == null)
+			{
+				ModelState.AddModelError("WhiteSpaces",
+						"Fields should not be white spaces alone.");
+			}
+
 			if (ModelState.IsValid)
 			{
+				// Deatach EF Core Entity from tracking same primary key value
+				_context.Entry(drawingFileBeforeEdit!).State = EntityState.Detached;
+
 				try
 				{
 					_context.Update(drawingFile);
@@ -131,19 +231,19 @@ namespace DrawingRegisterWeb.Controllers
 					{
 						return NotFound();
 					}
-					else
-					{
-						throw;
-					}
+					else throw;
 				}
 				return RedirectToAction("Details", "Drawings", new { id = drawingFile.DrawingId });
 			}
-			ViewData["DrawingId"] = new SelectList(_context.Drawing, "Id", "DrawingNumber", drawingFile.DrawingId);
-			return View(drawingFile);
 
+			// Return aproved DrawingId
+			ViewData["DrawingId"] = drawingFile.DrawingId;
+			return View(drawingFile);
 		}
 
-		// GET: DrawingFiles/Delete
+
+
+
 		public async Task<IActionResult> Delete(int? id)
 		{
 			if (id == null || _context.DrawingFile == null)
@@ -151,18 +251,24 @@ namespace DrawingRegisterWeb.Controllers
 				return NotFound();
 			}
 
-			var layout = await _context.DrawingFile
+			var user = await _userManager.GetUserAsync(User);
+			var drawingRegisterUser = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(dr => dr.UserId == user.Id);
+			var drawingFile = await _context.DrawingFile
 				.Include(d => d.Drawing)
+				.Include(d => d.Drawing.Project)
+				.Include(d => d.Drawing.Project.ProjectState)
+				.Where(d => d.Drawing.Project.ProjectState!.DrawingRegisterId == drawingRegisterUser!.DrawingRegisterId)
 				.FirstOrDefaultAsync(d => d.Id == id);
-			if (layout == null)
+
+			// Check if drawing file exists and ensure that user access drawing file that is only in his DrawingRegister
+			if (drawingFile == null)
 			{
 				return NotFound();
 			}
 
-			return View(layout);
+			return View(drawingFile);
 		}
 
-		// POST: DrawingFiles/Delete
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> DeleteConfirmed(int id)
@@ -172,29 +278,62 @@ namespace DrawingRegisterWeb.Controllers
 				return Problem("Entity set 'DrawingRegisterContext.DrawingFile'  is null.");
 			}
 
-			var drawingFile = await _context.DrawingFile.FindAsync(id);
-			int drawingId = drawingFile!.DrawingId;
+			var user = await _userManager.GetUserAsync(User);
+			var drawingRegisterUser = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(dr => dr.UserId == user.Id);
+			var drawingFile = await _context.DrawingFile
+				.Include(d => d.Drawing)
+				.Include(d => d.Drawing.Project)
+				.Include(d => d.Drawing.Project.ProjectState)
+				.Where(d => d.Drawing.Project.ProjectState!.DrawingRegisterId == drawingRegisterUser!.DrawingRegisterId)
+				.FirstOrDefaultAsync(d => d.Id == id);
 
+			// Check if drawing file exists and ensure that user access drawing file that is only in his DrawingRegister
 			if (drawingFile != null)
 			{
-				_context.DrawingFile.Remove(drawingFile);
-
-				var oldFilePath = Path.Combine(_hostEnvironment.WebRootPath, drawingFile.FileUrl.TrimStart('\\'));
-				int oldFileEndIndex = oldFilePath.LastIndexOf(".");
-				string thumbanilUrl = oldFilePath.Substring(0, oldFileEndIndex);
-				thumbanilUrl += ".jpg";
-
-				if (System.IO.File.Exists(oldFilePath) && !oldFilePath.Contains("SeededData"))
+				// Make sure that only administrator can edit if ProjectState is Defined, Completed or Canceled
+				if (drawingRegisterUser!.Role != ConstData.Role_Admin_Name)
 				{
-						System.IO.File.Delete(oldFilePath);
-						System.IO.File.Delete(thumbanilUrl);
+					if (drawingFile!.Drawing.Project.ProjectState!.Name == ConstData.State_Defined ||
+						drawingFile!.Drawing.Project.ProjectState!.Name == ConstData.State_Completed ||
+						drawingFile!.Drawing.Project.ProjectState!.Name == ConstData.State_Canceled)
+					{
+						TempData["NoDelete"] = $"Only the administrator has the ability to edit this drawing file " +
+							$"if project state is set to {drawingFile!.Drawing.Project.ProjectState!.Name}.";
+
+						return View(drawingFile);
+					}
 				}
 
-			}
+				// Delete old file and thumbnail if it exsist
+				var oldFilePath = Path.Combine(_hostEnvironment.WebRootPath, drawingFile.FileUrl.TrimStart('\\'));
 
-			await _context.SaveChangesAsync();
-			return RedirectToAction("Details", "Drawings", new { id = drawingId });
+				if (System.IO.File.Exists(oldFilePath) && !oldFilePath.Contains("SeededData")) // Prevent from deleting SeededData files
+				{
+					System.IO.File.Delete(oldFilePath);
+
+					if(drawingFile.FileType == "pdf")
+					{
+						int oldFileEndIndex = oldFilePath.LastIndexOf(".");
+						string thumbanilUrl = oldFilePath.Substring(0, oldFileEndIndex);
+
+						thumbanilUrl += ".jpg";
+
+						System.IO.File.Delete(thumbanilUrl);
+					}
+				}
+
+				_context.DrawingFile.Remove(drawingFile);
+
+				await _context.SaveChangesAsync();
+
+				return RedirectToAction("Details", "Drawings", new { id = drawingFile!.DrawingId });
+
+			} else return NotFound();
 		}
+
+
+
+
 		private bool DrawingFileExists(int id)
 		{
 			return _context.DrawingFile.Any(e => e.Id == id);
